workaround multinomial() int
Per #1399, In Numpy 1.26, `RandomState.multinomial()` rejects float value n. There could be more such API changes.

I'm asking if this is a bug and if it's a side effect of NumPy bindings moving from the Cython project to the NumPy project or by semantic changes in Cython 3.0.

+++ models/ecoli/processes/rna_degradation.py
@@ -347,8 +347,8 @@ class RnaDegradation(wholecell.processes.process.Process):
 		n_fragment_bases = self.fragment_bases.counts()
 		n_fragment_bases_sum = n_fragment_bases.sum()
 
-		exornase_capacity = n_exoRNases.sum() * self.kcat_exoRNase * (
-				units.s * self.timeStepSec())
+		exornase_capacity = int(n_exoRNases.sum() * self.kcat_exoRNase * (
+				units.s * self.timeStepSec()))
 
 		if exornase_capacity >= n_fragment_bases_sum:
 			self.nmps.countsInc(n_fragment_bases)
@@ -420,7 +420,7 @@ class RnaDegradation(wholecell.processes.process.Process):
 		while n_rnas_to_degrade.sum() < n_total_rnas_to_degrade and remaining_rna_counts.sum() != 0:
 			n_rnas_to_degrade += np.fmin(
 				self.randomState.multinomial(
-					n_total_rnas_to_degrade - n_rnas_to_degrade.sum(),
+					int(n_total_rnas_to_degrade - n_rnas_to_degrade.sum()),
 					rna_deg_probs
 					),
 				remaining_rna_counts
